+++ source_handlers.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/RedHatInsights/sources-api-go/dao"
 	m "github.com/RedHatInsights/sources-api-go/model"
+	"github.com/RedHatInsights/sources-api-go/service"
 	"github.com/RedHatInsights/sources-api-go/util"
 	"github.com/labstack/echo/v4"
 )
@@ -176,7 +177,7 @@ func SourceCreate(c echo.Context) error {
 		return err
 	}
 
-	err = input.Validate()
+	err = service.ValidateSourceCreationRequest(input)
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, util.ErrorDoc(fmt.Sprintf("Validation failed: %s", err.Error()), "400"))
 	}
